* fixed size_t -> int conversion warnings (Colin)

+++ workbench/ServoWorkbench/CCrystalTextBuffer.cpp
@@ -143,7 +143,7 @@ void CCrystalTextBuffer::InsertLine(LPCTSTR pszLine, int nLength /*= -1*/, int n
 		m_aLines.InsertAt(nPosition, li);
 
 #ifdef _DEBUG
-	int nLines = m_aLines.GetSize();
+	int nLines = (int)m_aLines.GetSize();
 	if (nLines % 5000 == 0)
 		TRACE1("%d lines loaded!\n", nLines);
 #endif
@@ -181,7 +181,7 @@ void CCrystalTextBuffer::AppendLine(int nLineIndex, LPCTSTR pszChars, int nLengt
 void CCrystalTextBuffer::FreeAll()
 {
 	//	Free text
-	int nCount = m_aLines.GetSize();
+	int nCount = (int)m_aLines.GetSize();
 	for (int I = 0; I < nCount; I ++)
 	{
 		if (m_aLines[I].m_nMax > 0)
@@ -190,7 +190,7 @@ void CCrystalTextBuffer::FreeAll()
 	m_aLines.RemoveAll();
 
 	//	Free undo buffer
-	int nBufSize = m_aUndoBuf.GetSize();
+	int nBufSize = (int)m_aUndoBuf.GetSize();
 	for (I = 0; I < nBufSize; I ++)
 		delete m_aUndoBuf[I].m_pcText;
 	m_aUndoBuf.RemoveAll();
@@ -406,9 +406,9 @@ BOOL CCrystalTextBuffer::SaveToFile(LPCTSTR pszFileName, int nCrlfStyle /*= CRLF
 
 		ASSERT(nCrlfStyle >= 0 && nCrlfStyle <= 2);
 		const char *pszCRLF = crlfs[nCrlfStyle];
-		int nCRLFLength = strlen(pszCRLF);
+		int nCRLFLength = (int)strlen(pszCRLF);
 
-		int nLineCount = m_aLines.GetSize();
+		int nLineCount = (int)m_aLines.GetSize();
 		USES_CONVERSION;
 		for (int nLine = 0; nLine < nLineCount; nLine ++)
 		{
@@ -490,7 +490,7 @@ int CCrystalTextBuffer::GetLineCount()
 {
 	ASSERT(m_bInit);	//	Text buffer not yet initialized.
 						//	You must call InitNew() or LoadFromFile() first!
-	return m_aLines.GetSize();
+	return (int)m_aLines.GetSize();
 }
 
 int CCrystalTextBuffer::GetLineLength(int nLine)
@@ -533,7 +533,7 @@ static int FlagToIndex(DWORD dwFlag)
 
 int CCrystalTextBuffer::FindLineWithFlag(DWORD dwFlag)
 {
-	int nSize = m_aLines.GetSize();
+	int nSize = (int)m_aLines.GetSize();
 	for (int L = 0; L < nSize; L ++)
 	{
 		if ((m_aLines[L].m_dwFlags & dwFlag) != 0)
@@ -964,7 +964,7 @@ void CCrystalTextBuffer::AddUndoRecord(BOOL bInsert, const CPoint &ptStartPos, c
 	ASSERT(m_aUndoBuf.GetSize() == 0 || (m_aUndoBuf[0].m_dwFlags & UNDO_BEGINGROUP) != 0);
 
 	//	Strip unnecessary undo records (edit after undo)
-	int nBufSize = m_aUndoBuf.GetSize();
+	int nBufSize = (int)m_aUndoBuf.GetSize();
 	if (m_nUndoPosition < nBufSize)
 	{
 		for (int I = m_nUndoPosition; I < nBufSize; I ++)
@@ -974,7 +974,7 @@ void CCrystalTextBuffer::AddUndoRecord(BOOL bInsert, const CPoint &ptStartPos, c
 
 	//	If undo buffer size is close to critical, remove the oldest records
 	ASSERT(m_aUndoBuf.GetSize() <= m_nUndoBufSize);
-	nBufSize = m_aUndoBuf.GetSize();
+	nBufSize = (int)m_aUndoBuf.GetSize();
 	if (nBufSize >= m_nUndoBufSize)
 	{
 		int nIndex = 0;
@@ -1016,7 +1016,7 @@ void CCrystalTextBuffer::AddUndoRecord(BOOL bInsert, const CPoint &ptStartPos, c
 		ur.m_pcText = NULL;
 
 	m_aUndoBuf.Add(ur);
-	m_nUndoPosition = m_aUndoBuf.GetSize();
+	m_nUndoPosition = (int)m_aUndoBuf.GetSize();
 
 	ASSERT(m_aUndoBuf.GetSize() <= m_nUndoBufSize);
 }
